GitHub Actions: refactoring package
