+++ src/main/java/tech/jhipster/lite/generator/ci/github/actions/domain/GithubActionsService.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.githubactions.domain;
+package tech.jhipster.lite.generator.ci.github.actions.domain;
 
 import tech.jhipster.lite.generator.project.domain.Project;
