+++ src/main/java/tech/jhipster/lite/generator/ci/github/actions/infrastructure/config/GithubActionsBeanConfiguration.java
@@ -1,9 +1,9 @@
-package tech.jhipster.lite.generator.githubactions.infrastructure.config;
+package tech.jhipster.lite.generator.ci.github.actions.infrastructure.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import tech.jhipster.lite.generator.githubactions.domain.GithubActionsDomainService;
-import tech.jhipster.lite.generator.githubactions.domain.GithubActionsService;
+import tech.jhipster.lite.generator.ci.github.actions.domain.GithubActionsDomainService;
+import tech.jhipster.lite.generator.ci.github.actions.domain.GithubActionsService;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 
 @Configuration
