+++ src/main/java/tech/jhipster/lite/generator/ci/github/actions/infrastructure/primary/rest/GithubActionsResource.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.githubactions.infrastructure.primary.rest;
+package tech.jhipster.lite.generator.ci.github.actions.infrastructure.primary.rest;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import tech.jhipster.lite.generator.githubactions.application.GithubActionsApplicationService;
+import tech.jhipster.lite.generator.ci.github.actions.application.GithubActionsApplicationService;
 import tech.jhipster.lite.generator.project.domain.Project;
 import tech.jhipster.lite.generator.project.infrastructure.primary.dto.ProjectDTO;
 import tech.jhipster.lite.technical.infrastructure.primary.annotation.GeneratorStep;
