+++ src/test/java/tech/jhipster/lite/generator/ci/github/actions/infrastructure/config/GithubActionsBeanConfigurationIT.java
@@ -1,14 +1,12 @@
-package tech.jhipster.lite.generator.githubactions.infrastructure.config;
+package tech.jhipster.lite.generator.ci.github.actions.infrastructure.config;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import tech.jhipster.lite.IntegrationTest;
-import tech.jhipster.lite.generator.githubactions.domain.GithubActionsDomainService;
-import tech.jhipster.lite.generator.init.domain.InitDomainService;
+import tech.jhipster.lite.generator.ci.github.actions.domain.GithubActionsDomainService;
 
 @IntegrationTest
 class GithubActionsBeanConfigurationIT {
